requirements.txt
fastapi
uvicorn
requests
python-dotenv
matplotlib
numpy